requirements.txt
requests
feedparser
beautifulsoup4
python-dateutil
google-generativeai
google-genai
pillow
